backend/src/requirements.txt
Flask==2.2.2
MarkupSafe==2.1.2
Jinja2==3.1.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
idna==2.10
requests==2.28.2
urllib3==1.26.3
waitress==1.4.4
Werkzeug==2.2.2
beautifulsoup4==4.8.0
pytz==2023.3
biplist==1.0.3
pycparser==2.19
six==1.16.0
soupsieve==1.9.3
#webapp2==2.5.2
#webauthn==1.8.0
lxml==4.9.2
cffi==1.15.1
#IPy==1.01
pkcs7==0.1.2
python-dateutil==2.8.2
pycryptodome==3.17
#pymysql
#flask_sqlalchemy
#flask_restful
#firebase-admin
appengine-python-standard==1.1.2
google-cloud-firestore
google-cloud-logging
google-cloud-storage
# google-cloud-datastore==2.15.1
google-api-python-client==2.157.0
apiclient==1.0.4
httplib2
oauth2client
oauth2
oauthlib==3.2.2
itsdangerous==2.1.2
signxml==3.1.1
feedparser==6.0.10
#elasticsearch>=7.17.9,<8.0.0
pyyaml==6.0
pyfcm==1.5.4
#xlsx2html==0.4.4
httpagentparser==1.9.5
#pyconfig==3.2.3
#apply==1.6
#memory-profiler==0.61.0
gunicorn==20.1
gspread==5.9.0
tinytag==1.9.0

psutil==5.9.5       # ログ出力で使用
pyasn1==0.5.0  # GAEGEN2対応：SAML、SSO連携対応


google-cloud-ndb
redis